[UHYU-90] feat: user 엔티티 수정에 따른 테스트 코드 일부 수정

+++ src/test/java/com/ureca/uhyu/domain/map/service/MapServiceImplTest.java
@@ -96,7 +96,7 @@ class MapServiceImplTest {
                 .userName("홍길동")
                 .kakaoId(456465L)
                 .email("asdad@kakao.com")
-                .age((byte) 32)
+                .age(32)
                 .gender(Gender.MALE)
                 .role(UserRole.TMP_USER)
                 .status(Status.ACTIVE)
